refactor: group e2e tests by product namespaces
