+++ Transbank.Tests/E2E/POSAutoservicio/POSAutoservicioLastSaleE2ETests.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 using Transbank.Responses.AutoservicioResponse;
 using Xunit;
 
-namespace Transbank.Tests.E2E
+namespace Transbank.Tests.E2E.POSAutoservicio
 {
     public class POSAutoservicioLastSaleE2ETests : POSAutoservicioE2ETestBase
     {
