+++ Transbank.Tests/E2E/POSAutoservicio/POSAutoservicioMultiCodeSaleE2ETests.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using Transbank.Responses.AutoservicioResponse;
 using Xunit;
 
-namespace Transbank.Tests.E2E
+namespace Transbank.Tests.E2E.POSAutoservicio
 {
     public class POSAutoservicioMultiCodeSaleE2ETests : POSAutoservicioE2ETestBase
     {
