+++ Transbank.Tests/E2E/POSIntegrado/POSIntegradoMultiCodeSaleE2ETests.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using Transbank.Responses.IntegradoResponses;
 using Xunit;
 
-namespace Transbank.Tests.E2E
+namespace Transbank.Tests.E2E.POSIntegrado
 {
     public class POSIntegradoMultiCodeSaleE2ETests : POSIntegradoE2ETestBase
     {
